+++ packages/llamaindex/src/storage/vectorStore/MongoDBAtlasVectorStore.ts
@@ -5,7 +5,10 @@ import { getEnv } from "@llamaindex/env";
 import type { BulkWriteOptions, Collection } from "mongodb";
 import { MongoClient } from "mongodb";
 import {
+  FilterCondition,
   VectorStoreBase,
+  type FilterOperator,
+  type MetadataFilter,
   type MetadataFilters,
   type VectorStoreNoEmbedModel,
   type VectorStoreQuery,
@@ -13,15 +16,51 @@ import {
 } from "./types.js";
 import { metadataDictToNode, nodeToMetadata } from "./utils.js";
 
-// Utility function to convert metadata filters to MongoDB filter
-function toMongoDBFilter(
-  standardFilters: MetadataFilters,
-): Record<string, any> {
-  const filters: Record<string, any> = {};
-  for (const filter of standardFilters?.filters ?? []) {
-    filters[filter.key] = filter.value;
+// define your Atlas Search index. See detail https://www.mongodb.com/docs/atlas/atlas-search/field-types/knn-vector/
+const DEFAULT_EMBEDDING_DEFINITION = {
+  type: "knnVector",
+  dimensions: 1536,
+  similarity: "cosine",
+};
+
+function mapLcMqlFilterOperators(operator: string): string {
+  const operatorMap: { [key in FilterOperator]?: string } = {
+    "==": "$eq",
+    "<": "$lt",
+    "<=": "$lte",
+    ">": "$gt",
+    ">=": "$gte",
+    "!=": "$ne",
+    in: "$in",
+    nin: "$nin",
+  };
+  const mqlOperator = operatorMap[operator as FilterOperator];
+  if (!mqlOperator) throw new Error(`Unsupported operator: ${operator}`);
+  return mqlOperator;
+}
+
+function toMongoDBFilter(filters?: MetadataFilters): Record<string, any> {
+  if (!filters) return {};
+
+  const createFilterObject = (mf: MetadataFilter) => ({
+    [mf.key]: {
+      [mapLcMqlFilterOperators(mf.operator)]: mf.value,
+    },
+  });
+
+  if (filters.filters.length === 1) {
+    return createFilterObject(filters.filters[0]);
   }
-  return filters;
+
+  if (filters.condition === FilterCondition.AND) {
+    return { $and: filters.filters.map(createFilterObject) };
+  }
+
+  if (filters.condition === FilterCondition.OR) {
+    return { $or: filters.filters.map(createFilterObject) };
+  }
+
+  throw new Error("filters condition not recognized. Must be AND or OR");
 }
 
 /**
@@ -34,6 +73,12 @@ export class MongoDBAtlasVectorSearch
 {
   storesText: boolean = true;
   flatMetadata: boolean = true;
+
+  dbName: string;
+  collectionName: string;
+  autoCreateIndex: boolean;
+  embeddingDefinition: Record<string, unknown>;
+  indexedMetadataFields: string[];
 
   /**
    * The used MongoClient. If not given, a new MongoClient is created based on the MONGODB_URI env variable.
@@ -92,13 +137,16 @@ export class MongoDBAtlasVectorSearch
    * Default: query.similarityTopK * 10
    */
   numCandidates: (query: VectorStoreQuery) => number;
-  private collection: Collection;
+  private collection?: Collection;
 
   constructor(
     init: Partial<MongoDBAtlasVectorSearch> & {
       dbName: string;
       collectionName: string;
       embedModel?: BaseEmbedding;
+      autoCreateIndex?: boolean;
+      indexedMetadataFields?: string[];
+      embeddingDefinition?: Record<string, unknown>;
     },
   ) {
     super(init.embedModel);
@@ -114,9 +162,14 @@ export class MongoDBAtlasVectorSearch
       this.mongodbClient = new MongoClient(mongoUri);
     }
 
-    this.collection = this.mongodbClient
-      .db(init.dbName ?? "default_db")
-      .collection(init.collectionName ?? "default_collection");
+    this.dbName = init.dbName ?? "default_db";
+    this.collectionName = init.collectionName ?? "default_collection";
+    this.autoCreateIndex = init.autoCreateIndex ?? true;
+    this.indexedMetadataFields = init.indexedMetadataFields ?? [];
+    this.embeddingDefinition = {
+      ...DEFAULT_EMBEDDING_DEFINITION,
+      ...(init.embeddingDefinition ?? {}),
+    };
     this.indexName = init.indexName ?? "default";
     this.embeddingKey = init.embeddingKey ?? "embedding";
     this.idKey = init.idKey ?? "id";
@@ -125,6 +178,43 @@ export class MongoDBAtlasVectorSearch
     this.numCandidates =
       init.numCandidates ?? ((query) => query.similarityTopK * 10);
     this.insertOptions = init.insertOptions;
+  }
+
+  async ensureCollection() {
+    if (!this.collection) {
+      const collection = await this.mongodbClient
+        .db(this.dbName)
+        .createCollection(this.collectionName);
+
+      this.collection = collection;
+    }
+
+    if (this.autoCreateIndex) {
+      const searchIndexes = await this.collection.listSearchIndexes().toArray();
+      const indexExists = searchIndexes.some(
+        (index) => index.name === this.indexName,
+      );
+      if (!indexExists) {
+        const additionalDefinition: Record<string, { type: string }> = {};
+        this.indexedMetadataFields.forEach((field) => {
+          additionalDefinition[field] = { type: "token" };
+        });
+        await this.collection.createSearchIndex({
+          name: this.indexName,
+          definition: {
+            mappings: {
+              dynamic: true,
+              fields: {
+                embedding: this.embeddingDefinition,
+                ...additionalDefinition,
+              },
+            },
+          },
+        });
+      }
+    }
+
+    return this.collection;
   }
 
   /**
@@ -145,20 +235,26 @@ export class MongoDBAtlasVectorSearch
         this.flatMetadata,
       );
 
+      // Include the specified metadata fields in the top level of the document (to help filter)
+      const populatedMetadata: Record<string, unknown> = {};
+      for (const field of this.indexedMetadataFields) {
+        populatedMetadata[field] = metadata[field];
+      }
+
       return {
         [this.idKey]: node.id_,
         [this.embeddingKey]: node.getEmbedding(),
         [this.textKey]: node.getContent(MetadataMode.NONE) || "",
         [this.metadataKey]: metadata,
+        ...populatedMetadata,
       };
     });
 
-    console.debug("Inserting data into MongoDB: ", dataToInsert);
-    const insertResult = await this.collection.insertMany(
+    const collection = await this.ensureCollection();
+    const insertResult = await collection.insertMany(
       dataToInsert,
       this.insertOptions,
     );
-    console.debug("Result of insert: ", insertResult);
     return nodes.map((node) => node.id_);
   }
 
@@ -169,7 +265,8 @@ export class MongoDBAtlasVectorSearch
    * @param deleteOptions Options to pass to the deleteOne function
    */
   async delete(refDocId: string, deleteOptions?: any): Promise<void> {
-    await this.collection.deleteMany(
+    const collection = await this.ensureCollection();
+    await collection.deleteMany(
       {
         [`${this.metadataKey}.ref_doc_id`]: refDocId,
       },
@@ -214,8 +311,8 @@ export class MongoDBAtlasVectorSearch
       },
     ];
 
-    console.debug("Running query pipeline: ", pipeline);
-    const cursor = await this.collection.aggregate(pipeline);
+    const collection = await this.ensureCollection();
+    const cursor = await collection.aggregate(pipeline);
 
     const nodes: BaseNode[] = [];
     const ids: string[] = [];
@@ -241,7 +338,6 @@ export class MongoDBAtlasVectorSearch
       ids,
     };
 
-    console.debug("Result of query (ids):", ids);
     return result;
   }
 }