+++ packages/autotool/examples/02_nextjs/CHANGELOG.md
@@ -1,5 +1,39 @@
 # @llamaindex/autotool-02-next-example
 
+## 0.1.45
+
+### Patch Changes
+
+- Updated dependencies [d9d6c56]
+- Updated dependencies [22ff486]
+- Updated dependencies [eed0b04]
+  - llamaindex@0.5.20
+  - @llamaindex/autotool@2.0.1
+
+## 0.1.44
+
+### Patch Changes
+
+- Updated dependencies [fcbf183]
+  - llamaindex@0.5.19
+  - @llamaindex/autotool@2.0.1
+
+## 0.1.43
+
+### Patch Changes
+
+- Updated dependencies [8b66cf4]
+  - llamaindex@0.5.18
+  - @llamaindex/autotool@2.0.1
+
+## 0.1.42
+
+### Patch Changes
+
+- Updated dependencies [c654398]
+  - llamaindex@0.5.17
+  - @llamaindex/autotool@2.0.1
+
 ## 0.1.41
 
 ### Patch Changes